# next-rms-app/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SmartRent API - Tài liệu API cho Mobile App",
    "version": "1.0.0",
    "description": "Tài liệu mô tả toàn bộ các API mà Mobile App (cho Manager và Tenant) cần kết nối. Tất cả các API cần xác thực đều phải gửi kèm `Authorization: Bearer <access_token>` trong header.\n\n## Cách lấy Access Token\nGọi API `POST /api/auth/login` với số điện thoại và mật khẩu. Token sẽ được trả về trong trường `access_token`.\n\n## Phân quyền (Roles)\n- **super_admin**: Toàn quyền.\n- **manager**: Chỉ xem/thao tác dữ liệu thuộc chi nhánh được gán.\n- **tenant**: Chỉ xem/thao tác dữ liệu của phòng mình đang thuê."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Môi trường phát triển (Local)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập Access Token (lấy từ API /api/auth/login)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Unauthorized" }
        }
      },
...
      "Room": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "roomCode": { "type": "string", "example": "101" },
          "floor": { "type": "integer", "example": 1 },
          "area": { "type": "number", "example": 25.0 },
          "basePrice": { "type": "integer", "example": 3500000 },
          "electricPrice": { "type": "integer", "example": 4000 },
          "waterPrice": { "type": "integer", "example": 15000 },
          "status": { "type": "string", "enum": ["available", "occupied", "maintenance"], "example": "occupied" },
          "branch": { "type": "integer", "example": 1 }
        }
      },
      "Ticket": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 5 },
          "title": { "type": "string", "example": "Đèn phòng bị hỏng" },
          "description": { "type": "string", "example": "Đèn phòng ngủ không sáng, cần thay bóng gấp." },
          "status": { "type": "string", "enum": ["pending", "in-progress", "resolved"], "example": "pending" },
          "priority": { "type": "string", "enum": ["low", "medium", "high"], "example": "high" },
          "images": { "type": "array", "items": { "type": "string", "format": "uri" }, "example": ["https://res.cloudinary.com/..."] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "security": [],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["🔑 Xác thực (Auth)"],
        "summary": "Đăng nhập",
        "description": "Đăng nhập bằng số điện thoại và mật khẩu. Trả về `access_token` và `refresh_token` để dùng cho các API khác, cùng với thông tin `role` và `branch_id` của người dùng.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone", "password"],
                "properties": {
                  "phone": { "type": "string", "description": "Số điện thoại (định dạng 0xxx hoặc +84xxx)", "example": "0901234567" },
                  "password": { "type": "string", "format": "password", "description": "Mật khẩu (tối thiểu 6 ký tự)", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng nhập thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "access_token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." },
                    "refresh_token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "phone": { "type": "string" },
                        "role": { "type": "string", "enum": ["super_admin", "manager", "tenant"] },
                        "branch_id": { "type": "integer", "nullable": true },
                        "full_name": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Thiếu hoặc sai định dạng thông tin đầu vào", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Sai số điện thoại hoặc mật khẩu", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "tags": ["🔑 Xác thực (Auth)"],
        "summary": "Đổi mật khẩu",
        "description": "Đổi mật khẩu cho tài khoản hiện tại đang đăng nhập. Yêu cầu gửi `Authorization: Bearer <token>` trong header.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["new_password"],
                "properties": {
                  "new_password": { "type": "string", "format": "password", "description": "Mật khẩu mới (tối thiểu 6 ký tự)", "example": "newPassword456" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Đổi mật khẩu thành công", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string", "example": "Đổi mật khẩu thành công" } } } } } },
          "400": { "description": "Mật khẩu mới không hợp lệ (quá ngắn)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Chưa đăng nhập", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/rooms/list": {
      "get": {
        "tags": ["🏠 Phòng (Rooms)"],
        "summary": "Lấy danh sách phòng",
        "description": "Trả về danh sách phòng có phân trang và nhiều bộ lọc. **Phân quyền:**\n- **super_admin**: Xem tất cả hoặc lọc theo `branch_id`.\n- **manager**: Chỉ thấy phòng thuộc chi nhánh của mình (không cần truyền `branch_id`).\n- **tenant**: Bị chặn, không được xem danh sách.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "description": "Số trang (mặc định: 1)", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "description": "Số lượng mỗi trang (mặc định: 10, tối đa: 100)", "schema": { "type": "integer", "default": 10 } },
          { "name": "status", "in": "query", "description": "Lọc theo trạng thái phòng", "schema": { "type": "string", "enum": ["available", "occupied", "maintenance"] } },
          { "name": "branch_id", "in": "query", "description": "Lọc theo chi nhánh (chỉ super_admin dùng được)", "schema": { "type": "integer" } },
          { "name": "search", "in": "query", "description": "Tìm kiếm theo số phòng (hỗ trợ tìm kiếm mờ)", "schema": { "type": "string" } },
          { "name": "floor", "in": "query", "description": "Lọc theo tầng", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "docs": { "type": "array", "items": { "$ref": "#/components/schemas/Room" } },
                    "totalDocs": { "type": "integer", "example": 50 },
                    "limit": { "type": "integer", "example": 10 },
                    "page": { "type": "integer", "example": 1 },
                    "totalPages": { "type": "integer", "example": 5 }
                  }
                }
              }
            }
          },
          "403": { "description": "Không có quyền xem danh sách phòng (Tenant bị chặn)" }
        }
      }
    },
    "/api/tickets": {
      "get": {
        "tags": ["🔧 Yêu cầu Bảo trì (Tickets)"],
        "summary": "Lấy danh sách yêu cầu bảo trì",
        "description": "**Phân quyền:**\n- **super_admin**: Xem tất cả tickets.\n- **manager**: Chỉ thấy tickets thuộc chi nhánh của mình.\n- **tenant**: Chỉ thấy tickets của phòng mình đang thuê.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "roomId", "in": "query", "description": "Lọc theo ID phòng", "schema": { "type": "integer" } },
          { "name": "status", "in": "query", "description": "Lọc theo trạng thái", "schema": { "type": "string", "enum": ["pending", "in-progress", "resolved"] } }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Ticket" } }
                  }
                }
              }
            }
          },
          "401": { "description": "Chưa đăng nhập" }
        }
      },
      "post": {
        "tags": ["🔧 Yêu cầu Bảo trì (Tickets)"],
        "summary": "Gửi yêu cầu bảo trì mới",
        "description": "Tạo một yêu cầu bảo trì mới. Thường được gọi bởi **Tenant** khi phát hiện sự cố trong phòng.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["roomId", "title"],
                "properties": {
                  "roomId": { "type": "integer", "example": 1, "description": "ID phòng cần bảo trì" },
                  "tenantId": { "type": "integer", "example": 5, "description": "ID khách thuê gửi yêu cầu" },
                  "title": { "type": "string", "example": "Đèn phòng bị hỏng", "description": "Tóm tắt sự cố (tối đa 100 ký tự)" },
                  "description": { "type": "string", "example": "Đèn phòng ngủ không sáng." },
                  "priority": { "type": "string", "enum": ["low", "medium", "high"], "default": "medium" },
                  "images": { "type": "array", "items": { "type": "string", "format": "uri" }, "description": "Mảng các URL ảnh đã được tải lên Cloudinary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tạo ticket thành công", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "$ref": "#/components/schemas/Ticket" } } } } } },
          "400": { "description": "Thiếu thông tin bắt buộc hoặc sai định dạng" },
          "401": { "description": "Chưa đăng nhập" }
        }
      }
    },
...
  }
}